+++ tests/configuration_manager/test_configuration_manager.py
@@ -236,6 +236,7 @@ def create_cli_context(tmpdir, app_version: str = "0.0.0") -> CliContext:
         docker_credentials_file=None,
         subcommand_args=None,
         debug=True,
+        is_dev_mode=False,
         app_name_slug=APP_NAME_SLUG,
         app_version=app_version,
         commands={},